delete routes folder, add addJob function
