+++ main.js
@@ -1,5 +1,6 @@
 import "./style.css";
-import "./src/index.js";
+import "./src/components/header.js";
+import "./src/components/footer.js";
 import CucumberRouter from "../cucumber-router/src/index.js";
 import "cucumber-components/src/components/side-nav/side-nav.js";
 import "cucumber-components/src/components/side-nav-item/side-nav-item.js";
@@ -15,36 +16,10 @@ import "cucumber-components/src/components/dialog/dialog.js";
 import "cucumber-components/src/components/button/button.js";
 import "cucumber-components/src/components/spinner/spinner.js";
 
+import "./src/controllers/job.js";
+
 const outlet = document.querySelector('[router-outlet]');
 const ROUTER = new CucumberRouter(outlet);
 window.ROUTER = ROUTER;
 
-
 ROUTER.start();
-
-// ROUTER.add('/jobs', async ({ route, params }) => {
-//   await import("./src/routes/jobs-list.js").catch(console.error);
-//   main.innerHTML = `<jobs-list></jobs-list>`;
-//   // https://developer.mozilla.org/en-US/docs/Web/API/CustomElementRegistry/getName
-//   // not supported yet
-//   // console.log(customElements.getName(signup));
-// });
-
-// ROUTER.add('/jobs/edit', async ({ route, params }) => {
-//   await import("./src/routes/edit-job.js").catch(console.error);
-//   const editJob = document.createElement('edit-job');
-//   main.appendChild(editJob);
-// });
-
-// ROUTER.add('/jobs/delete', async ({ route, params }) => {
-//   await import("./src/routes/delete-job.js").catch(console.error);
-//   const deleteJob = document.createElement('delete-job');
-//   main.appendChild(deleteJob);
-// });
-
-// ROUTER.add('404', () => {
-//   alert('page not found');
-//   ROUTER.redirect('/');
-// });
-
-// ROUTER.start();
