--- src/index.js
@@ -1,2 +1,0 @@
-import "./components/header.js";
-import "./components/footer.js";
